+++ package.json
@@ -51,6 +51,7 @@
     "autoprefixer": "^10.4.14",
     "eslint": "^8.47.0",
     "eslint-config-next": "^13.4.19",
+    "eslint-plugin-import": "^2.29.1",
     "jsdom": "^22.1.0",
     "postcss": "^8.4.27",
     "prettier": "^3.0.0",
