+++ src/features/food/components/Food/Food.test.tsx
@@ -1,7 +1,8 @@
-import { render, screen } from "~/utils/test/test-utils";
-import { Food } from "./Food";
 import { type Food as TFood } from "@prisma/client";
 import { test, expect, vi } from "vitest";
+
+import { render, screen } from "~/utils/test/test-utils";
+import { Food } from "./Food";
 
 const mockFoodData: TFood = {
   id: "1",
