+++ src/features/food/utils/calculateFreezerTime.ts
@@ -1,6 +1,7 @@
 import { FoodTypes } from "@prisma/client";
 import { formatDistanceToNowStrict, isFuture } from "date-fns";
 import { es } from "date-fns/locale";
+
 import { FREEZE_STATES } from "../components/Food/constants";
 
 type PausedState = {
