+++ src/features/space/components/Space.test.tsx
@@ -1,7 +1,8 @@
-import { render, screen } from "~/utils/test/test-utils";
-import { Space } from "./Space";
 import { type Space as TSpace } from "@prisma/client";
 import { test, expect } from "vitest";
+
+import { render, screen } from "~/utils/test/test-utils";
+import { Space } from "./Space";
 
 const mockSpaceData: TSpace = {
   id: "1548451516",
