+++ src/pages/_app.tsx
@@ -3,7 +3,6 @@ import { SessionProvider } from "next-auth/react";
 import { type AppType } from "next/app";
 
 import { api } from "~/utils/api";
-
 import "~/styles/globals.css";
 
 const MyApp: AppType<{ session: Session | null }> = ({
