+++ src/pages/create-space.tsx
@@ -1,11 +1,10 @@
 import { type FormEvent } from "react";
 import { useRouter } from "next/navigation";
 import Head from "next/head";
-
 import { Container as ContainerIcon } from "lucide-react";
 
 import { api } from "~/utils/api";
-import Container from "~/features/common/Container";
+import { Container } from "~/features/common";
 import { createSpace } from "~/utils/schemas/space";
 
 export default function CreateSpace() {
