+++ src/pages/space/[id]/join/[inv_id]/index.tsx
@@ -1,8 +1,8 @@
 import { signIn, useSession } from "next-auth/react";
 import { useRouter } from "next/router";
-import Container from "~/features/common/Container";
+
 import QueryErrorBoundary from "~/features/common/Error/QueryErrorBoundary";
-import Loader from "~/features/common/Loader";
+import { Loader, Container } from "~/features/common";
 import { api } from "~/utils/api";
 
 export default function JoinWithIvitation() {
