+++ src/server/api/routers/space.ts
@@ -6,7 +6,6 @@ import {
   joinWithInvitationRequest,
   setPrivacy,
 } from "~/utils/schemas/space";
-
 import {
   createTRPCRouter,
   protectedProcedure,
