+++ src/utils/api.ts
@@ -6,11 +6,7 @@
  */
 import { TRPCClientError, httpBatchLink, loggerLink } from "@trpc/client";
 import { createTRPCNext } from "@trpc/next";
-import {
-  TRPCError,
-  type inferRouterInputs,
-  type inferRouterOutputs,
-} from "@trpc/server";
+import type { inferRouterInputs, inferRouterOutputs } from "@trpc/server";
 import superjson from "superjson";
 
 import { type AppRouter } from "~/server/api/root";
