+++ src/utils/test/setup.ts
@@ -1,2 +1,3 @@
-import "~/styles/globals.css";
 import "@testing-library/jest-dom/vitest";
+
+import "~/styles/globals.css";
